requirements.txt
discord.py[voice]>=2.3.0
yt-dlp>=2023.7.6
pynacl>=1.5.0
spotipy==2.25.1